+++ ajedrez/all_pieces/king.py
@@ -1,24 +1,21 @@
-from ajedrez.pieces import Piece
+from ajedrez.pieces import Piece, Directions
 
-class King(Piece):
+class Queen(Piece):
     def __init__(self, color, row=None, col=None):
-        super().__init__(color, "♔", "♚", row, col)
+        super().__init__(color, "♕", "♛", row, col)
 
-    def basic_king_moves(self, row, col):
+    def basic_queen_moves(self, row, col):
         moves = []
-        directions = [
-            (-1, -1), (-1, 0), (-1, 1),
-            ( 0, -1),         ( 0, 1),
-            ( 1, -1), ( 1, 0), ( 1, 1)
-        ]
-
-        for direction_row, direction_col in directions:
-            r, c = row + direction_row, col + direction_col
-
-            if self.is_within_bounds(r, c):
+        directions = Directions.diagonal() + Directions.straight()  # Usar las direcciones
+        
+        for direction_r, direction_c in directions:
+            r, c = row + direction_r, col + direction_c
+            while 0 <= r < 8 and 0 <= c < 8:
                 moves.append((r, c))
-                
+                r += direction_r
+                c += direction_c
+
         return moves
     
     def get_possible_moves(self, board, row, col):
-        return self.basic_king_moves(row, col)
+        return self.basic_queen_moves(row, col)